Better code separation
